Put schema in a separate file
